clranup util imports (also magic strings) + remove ruamel dependency in subprocesses
