+++ amlb/utils/cache.py
@@ -4,7 +4,7 @@ from .core import flatten
 
 log = logging.getLogger(__name__)
 
-__no_export = dir()  # all variables defined above this are not exported
+__no_export = set(dir())  # all variables defined above this are not exported
 
 _CACHE_PROP_PREFIX_ = '__cached__'
