+++ amlb/utils/modules.py
@@ -1,4 +1,6 @@
 import logging
+import sys
+import types
 
 try:
     from pip._internal import main as pip_main
@@ -8,7 +10,20 @@ except ImportError:
 
 log = logging.getLogger(__name__)
 
-__no_export = dir()  # all variables defined above this are not exported
+__no_export = set(dir())  # all variables defined above this are not exported
+
+
+def register_module(module_name):
+    if module_name not in sys.modules:
+        mod = types.ModuleType(module_name)
+        sys.modules[module_name] = mod
+    return sys.modules[module_name]
+
+
+def register_submodule(mod, name):
+    fullname = '.'.join([mod.__name__, name])
+    module = register_module(fullname)
+    setattr(mod, name, module)
 
 
 def pip_install(module_or_requirements, is_requirements=False):
