+++ amlb/utils/os.py
@@ -12,7 +12,7 @@ from .time import datetime_iso
 
 log = logging.getLogger(__name__)
 
-__no_export = dir()  # all variables defined above this are not exported
+__no_export = set(dir())  # all variables defined above this are not exported
 
 
 def to_mb(size_in_bytes):
