+++ frameworks/shared/requirements.in
@@ -1,5 +1,4 @@
 psutil>=5.4
-ruamel.yaml>=0.15
 
 pyarrow>=4.0
 #tables>=3.6    # if using hdf serializer (let the client choose)
