+++ frameworks/shared/requirements.txt
@@ -1,5 +1,5 @@
 #
-# This file is autogenerated by pip-compile
+# This file is autogenerated by pip-compile with python 3.7
 # To update, run:
 #
 #    pip-compile frameworks/shared/requirements.in
@@ -9,8 +9,4 @@ numpy==1.21.0
 psutil==5.8.0
     # via -r frameworks/shared/requirements.in
 pyarrow==4.0.1
-    # via -r frameworks/shared/requirements.in
-ruamel.yaml.clib==0.2.2
-    # via ruamel.yaml
-ruamel.yaml==0.17.4
     # via -r frameworks/shared/requirements.in
